pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "continual-growing"
version = "0.1.0"
description = "Continual learning with growing Bayesian neural networks."
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "torch>=2.1",
    "torchvision>=0.16",
    "pyyaml>=6.0",
    "wandb>=0.16",
    "tqdm>=4.65",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
hpo = [
    "optuna>=3.5",
    "matplotlib>=3.7",
]

[project.scripts]
continual-growing-hpo = "hpo.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# CUDA 12.8 wheels for torch / torchvision. Required: NVIDIA driver R550+
# (nvidia-smi `CUDA Version` >= 12.8). For other setups override the
# index-url at install time, e.g. cpu-only:
#   pip install --index-url https://download.pytorch.org/whl/cpu torch torchvision
# These [[tool.uv.index]] entries are uv-only; pip users can ignore them
# and either rely on the default PyPI wheels or pass --index-url manually.
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
